+++ serialx/platforms/serial_pyodide/types.py
